+++ lib/rubocop/cop/lint/raise_exception.rb
@@ -16,12 +16,12 @@ module RuboCop
         MSG = 'Use `StandardError` over `Exception`.'
 
         def_node_matcher :exception?, <<~PATTERN
-          (send nil? ${:raise :fail} (const _ :Exception) ... )
+          (send nil? ${:raise :fail} (const {cbase nil?} :Exception) ... )
         PATTERN
 
         def_node_matcher :exception_new_with_message?, <<~PATTERN
           (send nil? ${:raise :fail}
-            (send (const _ :Exception) :new ... ))
+            (send (const {cbase nil?} :Exception) :new ... ))
         PATTERN
 
         def on_send(node)